Add support for decimal types

+++ samples/SampleApi/Contracts/CustomerSales.cs
@@ -3,5 +3,5 @@
 public record CustomerSales(
     int CustomerKey,
     string CustomerName,
-    float SalesAmount,
-    float TotalCost);
+    decimal SalesAmount,
+    decimal TotalCost);
